+++ resources/views/pages/script.blade.php
@@ -1,4 +1,3 @@
-{{-- TODO: add title based on script name :) --}}
-<x-layout.dashboard-layout title="Scenarista | ">
-    <h1>Script</h1>
+<x-layout.dashboard-layout :title="'Scenarista | ' . $script->name">
+    <pre>{!! $script->toJson(JSON_PRETTY_PRINT) !!}</pre>
 </x-layout.dashboard-layout>
